Sound Effects
Here are some sound effects I found that I think could work with our project. Some editing with them may be needed to fit it in, but I think it's a good base. I separated the files into two different zip folders based on the license used. the CC0 seems to allow completely free usage, while the Attribution requires we credit the author or something. All of the files were downloaded off of freesound.org and I included the licenses as found on the website with the zip folders.